Reorganizing non free build process.
- BUILD_NONFREE is OFF by default
