+++ src/backend/opencl/kernel/sift_nonfree.cl
