+++ src/backend/opencl/sift.cpp
@@ -14,7 +14,10 @@
 #include <Array.hpp>
 #include <err_opencl.hpp>
 #include <handle.hpp>
-#include <kernel/sift.hpp>
+
+#ifdef AF_NONFREE
+#include <kernel/sift_nonfree.hpp>
+#endif
 
 using af::dim4;
 using af::features;
@@ -30,6 +33,7 @@ unsigned sift(Array<float>& x_out, Array<float>& y_out, Array<float>& score_out,
               const float init_sigma, const bool double_input,
               const float img_scale, const float feature_ratio)
 {
+#ifdef AF_NONFREE
     unsigned nfeat_out;
     unsigned desc_len;
 
@@ -57,16 +61,19 @@ unsigned sift(Array<float>& x_out, Array<float>& y_out, Array<float>& score_out,
     }
 
     return nfeat_out;
+#else
+    AF_ERROR("ArrayFire was not built with nonfree support, SIFT disabled\n", AFF_ERR_NONFREE);
+#endif
 }
 
 
-#define INSTANTIATE(T, convAccT)                                                        \
+#define INSTANTIATE(T, convAccT)                                        \
     template unsigned sift<T, convAccT>(Array<float>& x_out, Array<float>& y_out, \
                                         Array<float>& score_out, Array<float>& ori_out, \
-                                        Array<float>& size_out, Array<float>& desc_out,  \
-                                        const Array<T>& in, const unsigned n_layers,                \
-                                        const float contrast_thr, const float edge_thr,             \
-                                        const float init_sigma, const bool double_input,            \
+                                        Array<float>& size_out, Array<float>& desc_out, \
+                                        const Array<T>& in, const unsigned n_layers, \
+                                        const float contrast_thr, const float edge_thr, \
+                                        const float init_sigma, const bool double_input, \
                                         const float img_scale, const float feature_ratio);
 
 INSTANTIATE(float , float )
