+++ test/sift_nonfree.cpp
